refactor: divide linting rules into each files
