+++ src/rules.rs
@@ -1,0 +1,2 @@
+pub mod compare_table_suffix_with_subquery;
+pub mod use_current_date;
